Add validação de erro para delete, update, find

+++ src/controllers/paletas.controller.js
@@ -7,7 +7,16 @@ const findPaletasController = (req, res) => {
 
 const findPaletaByIdController = (req, res) => {
     const idParam = req.params.id;
+    if (!idParam) {
+        return res.status(400).send({ message: 'ID não informado!' });
+    }
+
     const chosenPaleta = paletasService.findPaletaByIdService(idParam);
+
+    if (!chosenPaleta) {
+        return res.status(404).send({ message: 'Paleta não encontrada!' });
+    }
+
     res.send(chosenPaleta);
 };
 
@@ -32,6 +41,23 @@ const createPaletaController = (req, res) => {
 const updatePaletaController = (req, res) => {
     const idParam = +req.params.id;
     const paletaEdit = req.body;
+
+    if (!idParam) {
+        return res.status(404).send({ message: 'Paleta não encontrada!' });
+    }
+
+    if (
+        !paletaEdit ||
+        !paletaEdit.sabor ||
+        !paletaEdit.descricao ||
+        !paletaEdit.foto ||
+        !paletaEdit.preco
+    ) {
+        return res.status(400).send({
+            message: 'Você não preencheu todos os dados para editar a paleta!',
+        });
+    }
+
     const updatedPaleta = paletasService.updatePaletaService(
         idParam,
         paletaEdit,
@@ -41,6 +67,9 @@ const updatePaletaController = (req, res) => {
 
 const deletePaletaController = (req, res) => {
     const idParam = req.params.id;
+    if (!idParam) {
+        return res.status(404).send({ message: 'Paleta não encontrada!' });
+    }
     paletasService.deletePaletaService(idParam);
     res.send({ message: 'Paleta deletada com sucesso!' });
 };
